feat:매출 도메인 추가

+++ src/main/java/com/mcfuturepartners/crm/api/product/dto/ProductDto.java
@@ -14,9 +14,9 @@ import javax.validation.constraints.NotNull;
 @Data
 @NoArgsConstructor
 public class ProductDto {
-    @ApiModelProperty(position = 0, example = "상품 고유 id / 지금 필요 없음") private long id;
+    @ApiModelProperty(position = 0, example = "상품 고유 id / 지금 필요 없음") private Long id;
     @ApiModelProperty(position = 1, example = "상품명") private String name;
-    @ApiModelProperty(position = 2, example = "상품가격") private int price;
+    @ApiModelProperty(position = 2, example = "상품가격") private Long price;
 
     public Product toEntity(){
         return Product.builder()
