+++ src/main/java/com/mcfuturepartners/crm/api/user/dto/UserRevenueDto.java
@@ -17,13 +17,13 @@ public class UserRevenueDto {
                                 .getRegDate()
                                 .isAfter(LocalDate.now().atStartOfDay()))
                         .map(order -> order.getProduct().getPrice())
-                        .reduce(0, Integer::sum))
+                        .reduce(0L, Long::sum))
                 .monthlySales(user.getOrders().stream()
                         .filter(order -> order
                                 .getRegDate()
                                 .isAfter(LocalDate.of(LocalDate.now().getYear(),LocalDate.now().getMonth().getValue(),1).atStartOfDay()))
                         .map(order -> order.getProduct().getPrice())
-                        .reduce(0, Integer::sum))
+                        .reduce(0L, Long::sum))
                 .build();
     }
 }
